+++ CHANGES.md
@@ -2,6 +2,7 @@ v0.2.7 (dev)
 ======
 + new subcommand `ancestry` to predict ancestry using a simple neural network on the somalier
   sketches. creates an interactive html output and a text file
++ fix for "Argument list too long" on huge cohorts (#37)
 
 v0.2.6
 ======