+++ CHANGES.md
@@ -1,9 +1,10 @@
-v0.2.7 (dev)
+v0.2.7
 ======
 + new subcommand `ancestry` to predict ancestry using a simple neural network on the somalier
   sketches. creates an interactive html output and a text file
 + fix for "Argument list too long" on huge cohorts (#37)
 + sub-sample .pairs.tsv output for huge cohorts -- only for unrelated samples.
++ better sub-sampling of html output
 
 v0.2.6
 ======